shop -注意事項
1.更新的時候，可以把 application.properties檔案 取消打勾
2.必須自行新增商品

+++ workspace/springbootdemo/src/main/java/com/loveGod/demo/model/ProductDao.java
@@ -2,7 +2,7 @@ package com.loveGod.demo.model;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductDao extends JpaRepository<Products, Long> {
+public interface ProductDao extends JpaRepository<Products, Integer> {
 //	@Query(value = "from Products where name = :name")
 //	List<Products> findProductsByName(@Param("name")String name);
 //
